requirements.txt
# DispyCluster Requirements
# Python 3.9+ compatible, versions optimisées pour Raspberry Pi

# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.8.5
httpx==0.25.2
apscheduler==3.9.1
requests==2.31.0
pydantic==2.5.0
jinja2==3.1.2
python-multipart==0.0.6

# Dispy cluster framework
dispy==4.15.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
paramiko==3.2.0
cryptography==41.0.7

# Monitoring
prometheus-client==0.17.1
psutil==5.9.5

# Web scraping
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.10.0

# NLP et extraction d'entités
spacy==3.6.1
nltk==3.8.1
textblob==0.17.1

# Configuration et utilitaires
pyyaml==6.0.1
click==8.1.7
colorama==0.4.6
python-dotenv==1.0.0

# Celery et Redis
celery==5.3.6
redis==5.0.1

# Tests
pytest==7.4.2
pytest-asyncio==0.21.1
